+++ src/Practitioner/Practitioner/Practitioner.csproj
@@ -15,4 +15,8 @@
     <PackageReference Include="System.Text.Json" Version="4.7.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Views\Experience\" />
+  </ItemGroup>
+
 </Project>
