+++ src/Practitioner/Practitioner/Views/BoardsCommittees/BoardsCommitteesEdit.cshtml
@@ -1,6 +1,6 @@
 ﻿@model BoardsCommittees
 
-    <form asp-controller="Experience" asp-action="BoardsCommitteesEdit" method="post" role="form">
+    <form asp-controller="BoardsCommittees" asp-action="BoardsCommitteesEdit" method="post" role="form">
 
         <div asp-validation-summary="All" class="text-danger"></div>
