+++ src/Practitioner/Practitioner/Views/GeneralInterest/GeneralInterestEdit.cshtml
@@ -1,6 +1,6 @@
 ﻿@model GeneralInterest
 
-    <form asp-controller="Experience" asp-action="GeneralInterestEdit" method="post" role="form">
+    <form asp-controller="GeneralInterest" asp-action="GeneralInterestEdit" method="post" role="form">
 
         <div asp-validation-summary="All" class="text-danger"></div>
