+++ src/Practitioner/Practitioner/Views/ReloInterest/ReloInterestEdit.cshtml
@@ -1,6 +1,6 @@
 ﻿@model ReloInterest
 
-    <form asp-controller="Experience" asp-action="ReloInterestEdit" method="post" role="form">
+    <form asp-controller="ReloInterest" asp-action="ReloInterestEdit" method="post" role="form">
 
         <div asp-validation-summary="All" class="text-danger"></div>
