+++ src/Practitioner/Practitioner/Views/Shared/_AccountServedTable.cshtml
@@ -66,7 +66,7 @@
                                         </div>
                                         <div class="modal-footer">
                                             <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                                            <a class="btn btn-danger" asp-controller="Experience" asp-action="AccountServedDelete"
+                                            <a class="btn btn-danger" asp-controller="AccountServed" asp-action="AccountServedDelete"
                                                asp-route-id="@a.AccountServedId" asp-route-employeeId="@a.EmployeeId">Delete</a>
                                         </div>
                                     </div>
