+++ src/Practitioner/Practitioner/Views/Shared/_BoardsCommitteesTable.cshtml
@@ -66,7 +66,7 @@
                                         </div>
                                         <div class="modal-footer">
                                             <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                                            <a class="btn btn-danger" asp-controller="Experience" asp-action="BoardsCommitteesDelete"
+                                            <a class="btn btn-danger" asp-controller="BoardsCommittees" asp-action="BoardsCommitteesDelete"
                                                asp-route-id="@a.BoardsCommitteesId" asp-route-employeeId="@a.EmployeeId">Delete</a>
                                         </div>
                                     </div>
