+++ src/Practitioner/Practitioner/Views/Shared/_CareerInterestTable.cshtml
@@ -64,7 +64,7 @@
                                         </div>
                                         <div class="modal-footer">
                                             <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                                            <a class="btn btn-danger" asp-controller="Experience" asp-action="CareerInterestDelete"
+                                            <a class="btn btn-danger" asp-controller="CareerInterest" asp-action="CareerInterestDelete"
                                                asp-route-id="@a.CareerInterestId" asp-route-employeeId="@a.EmployeeId">Delete</a>
                                         </div>
                                     </div>
