+++ src/Practitioner/Practitioner/Views/Shared/_ExperienceTabsNav.cshtml
@@ -3,13 +3,13 @@
 
     <ul class="nav nav-tabs">
         <li class="nav-item">
-            <a class="nav-link @Model.AccountsServedActive" asp-controller="Experience" asp-action="AccountServed" asp-route-id=@Model.EmployeeId>Accounts Served</a>
+            <a class="nav-link @Model.AccountsServedActive" asp-controller="AccountServed" asp-action="AccountServed" asp-route-id=@Model.EmployeeId>Accounts Served</a>
         </li>
         <li class="nav-item">
-            <a class="nav-link @Model.PracticeManagementActive" asp-controller="Experience" asp-action="PracticeManagement" asp-route-id=@Model.EmployeeId>Practice Management</a>
+            <a class="nav-link @Model.PracticeManagementActive" asp-controller="PracticeManagement" asp-action="PracticeManagement" asp-route-id=@Model.EmployeeId>Practice Management</a>
         </li>
         <li class="nav-item">
-            <a class="nav-link @Model.BoardCommitteeActive" asp-controller="Experience" asp-action="BoardsCommittees" asp-route-id=@Model.EmployeeId>Boards and Committees</a>
+            <a class="nav-link @Model.BoardCommitteeActive" asp-controller="BoardsCommittees" asp-action="BoardsCommittees" asp-route-id=@Model.EmployeeId>Boards and Committees</a>
         </li>
         @*<li class="nav-item">
             <a class="nav-link @Model.TalentContributionActive" href="#">Talent Contributions</a>
