+++ src/Practitioner/Practitioner/Views/Shared/_GeneralInterestTable.cshtml
@@ -62,7 +62,7 @@
                                         </div>
                                         <div class="modal-footer">
                                             <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                                            <a class="btn btn-danger" asp-controller="Experience" asp-action="GeneralInterestDelete"
+                                            <a class="btn btn-danger" asp-controller="GeneralInterest" asp-action="GeneralInterestDelete"
                                                asp-route-id="@a.GeneralInterestId" asp-route-employeeId="@a.EmployeeId">Delete</a>
                                         </div>
                                     </div>
