+++ src/Practitioner/Practitioner/Views/Shared/_PracticeManagementTable.cshtml
@@ -66,7 +66,7 @@
                                         </div>
                                         <div class="modal-footer">
                                             <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                                            <a class="btn btn-danger" asp-controller="Experience" asp-action="PracticeManagementDelete"
+                                            <a class="btn btn-danger" asp-controller="PracticeManagement" asp-action="PracticeManagementDelete"
                                                asp-route-id="@a.PracticeManagementId" asp-route-employeeId="@a.EmployeeId">Delete</a>
                                         </div>
                                     </div>
