+++ src/Practitioner/Practitioner/Views/Shared/_ReloInterestTable.cshtml
@@ -66,7 +66,7 @@
                                     </div>
                                     <div class="modal-footer">
                                         <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                                        <a class="btn btn-danger" asp-controller="Experience" asp-action="ReloInterestDelete"
+                                        <a class="btn btn-danger" asp-controller="ReloInterest" asp-action="ReloInterestDelete"
                                            asp-route-id="@a.ReloInterestId" asp-route-employeeId="@a.EmployeeId">Delete</a>
                                     </div>
                                 </div>
